Lesson 8 - Fullscreen and Resizing
Previously all examples used a canvas with a fixed size (800x600). In this lesson, the project was modified to make the canvas fit the entire screen and resize accordingly when you change the size of your window. Aside from this, added the ability to toggle in and out of fullscreen by double clicking on the canvas. Also modified a couple of settings relating to pixel ratio for a better viewing experience (the best is 2).